+++ api/src/main/kotlin/io/github/llh4github/lotus/api/api/auth/UserApi.kt
@@ -1,7 +1,7 @@
 package io.github.llh4github.lotus.api.api.auth
 
 import io.github.llh4github.lotus.api.api.BaseApi
-import io.github.llh4github.lotus.api.service.UserService
+import io.github.llh4github.lotus.api.service.auth.UserService
 import io.github.llh4github.lotus.commons.JsonWrapper
 import io.github.llh4github.lotus.model.auth.dto.CreateUserInput
 import io.swagger.v3.oas.annotations.Operation
