+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/impl/AuthenticateServiceImpl.kt
@@ -1,22 +1,22 @@
-package io.github.llh4github.lotus.api.service.impl
+package io.github.llh4github.lotus.api.service.auth.impl
 
 import io.github.llh4github.lotus.api.dto.LoginParam
 import io.github.llh4github.lotus.api.dto.LoginResult
 import io.github.llh4github.lotus.api.dto.LogoutParam
 import io.github.llh4github.lotus.api.dto.UserAuthDetails
 import io.github.llh4github.lotus.api.security.UserAuthToken
-import io.github.llh4github.lotus.api.service.AuthService
-import io.github.llh4github.lotus.api.service.TokenService
-import io.github.llh4github.lotus.api.service.UserDetailsServiceImpl
+import io.github.llh4github.lotus.api.service.auth.AuthenticateService
+import io.github.llh4github.lotus.api.service.security.TokenService
+import io.github.llh4github.lotus.api.service.security.UserDetailsServiceImpl
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 
 @Service
-class AuthServiceImpl(
+class AuthenticateServiceImpl(
     private val tokenService: TokenService,
     private val userDetailsService: UserDetailsServiceImpl
-) : AuthService {
+) : AuthenticateService {
     private val logger = KotlinLogging.logger {}
     override fun login(param: LoginParam): LoginResult {
         val details = userDetailsService.loadUserByUsername(param.username) as UserAuthDetails
