+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/impl/RoleServiceImpl.kt
@@ -1,8 +1,8 @@
-package io.github.llh4github.lotus.api.service.impl
+package io.github.llh4github.lotus.api.service.auth.impl
 
 import io.github.llh4github.lotus.api.dao.RoleDao
 import io.github.llh4github.lotus.api.exceptions.RoleModuleException
-import io.github.llh4github.lotus.api.service.RoleService
+import io.github.llh4github.lotus.api.service.auth.RoleService
 import io.github.llh4github.lotus.model.auth.dto.RoleAddInput
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.stereotype.Service
